refactoring: 1) trade: item count type change from float to int 2) show dialog with error text, when username or email is already used.

+++ app/src/main/java/ru/wintrade/mvp/model/entity/Trade.kt
@@ -16,7 +16,7 @@ data class Trade(
     val orderStatus: String,
     val orderNum: String,
     val price: Float,
-    val count: Float,
+    val count: Int,
     val currency: String,
     val date: Date,
     val profitCount: String?,
